In-place numba optimizations
