+++ cvtModel/src/cvt_simulator/models/pulley/primary_pulley_flyweight.py
@@ -18,10 +18,35 @@ from cvt_simulator.constants.car_specs import (
     INITIAL_FLYWEIGHT_RADIUS,
 )
 from cvt_simulator.utils.system_state import SystemState
-from cvt_simulator.utils.numba_kernels import (
-    primary_flyweight_force_kernel,
-    max_torque_primary_kernel,
-)
+from cvt_simulator.utils.numba_utils import maybe_njit
+
+
+@maybe_njit(cache=True, fastmath=True)
+def _primary_flyweight_force_kernel(
+    flyweight_mass: float,
+    angular_velocity: float,
+    flyweight_radius: float,
+    ramp_slope: float,
+) -> tuple[float, float, float, float]:
+    centrifugal_force = flyweight_mass * angular_velocity**2 * flyweight_radius
+    angle = math.atan(-ramp_slope)
+    angle_multiplier = math.tan(angle)
+    net = centrifugal_force * angle_multiplier
+    return angle, centrifugal_force, angle_multiplier, net
+
+
+@maybe_njit(cache=True, fastmath=True)
+def _max_torque_primary_kernel(
+    mu_effective: float,
+    wrap_angle: float,
+    total_radial: float,
+    radius: float,
+) -> float:
+    exp_term = math.exp(mu_effective * wrap_angle)
+    capstan_term = (exp_term - 1.0) / (exp_term + 1.0)
+    radial_force_term = total_radial * radius / math.sin(wrap_angle / 2.0)
+    max_torque = capstan_term * radial_force_term
+    return max(0.0, max_torque)
 
 
 # TODO: Remove this code
@@ -156,7 +181,7 @@ class PhysicalPrimaryPulley(PrimaryPulleyModel):
         wrap_angle = self._get_wrap_angle(state.shift_distance)
         radius = self._get_radius(state.shift_distance)
 
-        return max_torque_primary_kernel(self.μ, wrap_angle, total_radial, radius)
+        return _max_torque_primary_kernel(self.μ, wrap_angle, total_radial, radius)
 
     # Private helper methods for force calculations
 
@@ -174,7 +199,7 @@ class PhysicalPrimaryPulley(PrimaryPulleyModel):
             shift_distance
         )
 
-        angle, centrifugal_force, angle_multiplier, net = primary_flyweight_force_kernel(
+        angle, centrifugal_force, angle_multiplier, net = _primary_flyweight_force_kernel(
             self.flyweight_mass,
             angular_velocity,
             flyweight_radius,
