+++ cvtModel/src/cvt_simulator/models/pulley/pulley_interface.py
@@ -35,7 +35,30 @@ from cvt_simulator.constants.constants import (
 )
 from cvt_simulator.models.dataTypes import PulleyState, PulleyForces, PulleyBreakdowns
 from cvt_simulator.utils.theoretical_models import TheoreticalModels as tm
-from cvt_simulator.utils.numba_kernels import radial_force_kernel
+from cvt_simulator.utils.numba_utils import maybe_njit
+
+
+@maybe_njit(cache=True, fastmath=True)
+def _radial_force_kernel(
+    clamping_force: float,
+    sheave_angle: float,
+    wrap_angle: float,
+    sec_angular_velocity: float,
+    sec_radius: float,
+    belt_cross_sectional_area: float,
+    rubber_density: float,
+) -> tuple[float, float, float]:
+    radial_from_clamping = 2.0 * (clamping_force * np.tan(sheave_angle / 2.0)) / wrap_angle
+
+    radial_from_centrifugal = (
+        sec_angular_velocity**2 * sec_radius**2 * belt_cross_sectional_area * rubber_density
+    )
+
+    total_radial = (
+        2.0 * np.sin(wrap_angle / 2.0) * (radial_from_clamping + radial_from_centrifugal)
+    )
+
+    return radial_from_clamping, radial_from_centrifugal, total_radial
 
 
 def get_kwarg(kwargs: dict[str, Any], key: str, default: Any = None) -> Any:
@@ -183,16 +206,14 @@ class PulleyModel(ABC):
         wheel_to_sec_ratio = GEARBOX_RATIO / WHEEL_RADIUS
         sec_angular_velocity = state.car_velocity * wheel_to_sec_ratio
 
-        radial_from_clamping, radial_from_centrifugal, total_radial = (
-            radial_force_kernel(
-                clamping_force,
-                SHEAVE_ANGLE,
-                wrap_angle,
-                sec_angular_velocity,
-                sec_radius,
-                BELT_CROSS_SECTIONAL_AREA,
-                RUBBER_DENSITY,
-            )
+        radial_from_clamping, radial_from_centrifugal, total_radial = _radial_force_kernel(
+            clamping_force,
+            SHEAVE_ANGLE,
+            wrap_angle,
+            sec_angular_velocity,
+            sec_radius,
+            BELT_CROSS_SECTIONAL_AREA,
+            RUBBER_DENSITY,
         )
 
         return radial_from_clamping, radial_from_centrifugal, total_radial
