+++ cvtModel/src/cvt_simulator/models/pulley/secondary_pulley_torque_reactive.py
@@ -16,10 +16,55 @@ from cvt_simulator.constants.car_specs import (
 )
 from cvt_simulator.models.ramps import LinearSegment, PiecewiseRamp
 from cvt_simulator.utils.system_state import SystemState
-from cvt_simulator.utils.numba_kernels import (
-    secondary_helix_force_kernel,
-    max_torque_secondary_kernel,
-)
+from cvt_simulator.utils.numba_utils import maybe_njit
+
+
+@maybe_njit(cache=True, fastmath=True)
+def _secondary_helix_force_kernel(
+    torque: float,
+    spring_torsion_torque: float,
+    secondary_radius: float,
+    ramp_slope: float,
+) -> tuple[float, float, float]:
+    helix_angle = np.arctan(-ramp_slope)
+    angle_multiplier = 2.0 * np.tan(helix_angle) * secondary_radius
+    net = (torque + spring_torsion_torque) / angle_multiplier
+    return helix_angle, angle_multiplier, net
+
+
+@maybe_njit(cache=True, fastmath=True)
+def _max_torque_secondary_kernel(
+    spring_comp_force: float,
+    spring_tors_torque: float,
+    sheave_angle: float,
+    radial_from_centrifugal: float,
+    wrap_angle: float,
+    radius: float,
+    mu_effective: float,
+    cvt_ratio: float,
+    helix_radius: float,
+    helix_angle: float,
+) -> float:
+    spring_force_term = (spring_comp_force + spring_tors_torque) * np.tan(
+        sheave_angle / 2.0
+    )
+
+    centrifugal_force = radial_from_centrifugal * wrap_angle / 2.0
+
+    exp_term = np.exp(mu_effective * wrap_angle)
+    capstan_term = (wrap_angle / (4.0 * radius)) * (exp_term + 1.0) / (exp_term - 1.0)
+
+    transmission_term = (
+        2.0
+        * cvt_ratio
+        * (helix_radius * np.tan(helix_angle))
+        * np.tan(sheave_angle / 2.0)
+    )
+
+    numerator = centrifugal_force + spring_force_term
+    denominator = capstan_term - transmission_term
+    max_torque = numerator / denominator
+    return max(0.0, max_torque)
 
 
 # TODO: Remove this code
@@ -87,6 +132,8 @@ class PhysicalSecondaryPulley(SecondaryPulleyModel):
         self.helix_radius = HELIX_RADIUS
         # This ramp needs to have a unique x for every height
         self.ramp = ramp
+        last_segment = self.ramp.segments[-1]
+        self._max_supported_shift = abs(self.ramp.height(last_segment.x_end))
 
     def calculate_clamping_force(
         self, state: SystemState, **kwargs
@@ -168,7 +215,7 @@ class PhysicalSecondaryPulley(SecondaryPulleyModel):
         # Torque transmission term (feedback loop)
         cvt_ratio = tm.current_cvt_ratio(shift_distance)
         helix_angle = helix_breakdown.angle
-        return max_torque_secondary_kernel(
+        return _max_torque_secondary_kernel(
             spring_comp_force,
             spring_tors_torque,
             SHEAVE_ANGLE,
@@ -193,7 +240,7 @@ class PhysicalSecondaryPulley(SecondaryPulleyModel):
         """
         # Clamp shift distance to valid range
         # TODO: Remove
-        shift_distance = np.clip(shift_distance, 0, MAX_SHIFT)
+        shift_distance = np.clip(shift_distance, 0, self._max_supported_shift)
 
         # Calculate torsion spring torque (resists cam rotation)
         spring_torque_breakdown = self._calculate_spring_tors_torque(shift_distance)
@@ -201,7 +248,7 @@ class PhysicalSecondaryPulley(SecondaryPulleyModel):
         # Effective radius at current shift position
         secondary_radius = tm.outer_sec_radius(shift_distance) - BELT_HEIGHT / 2
 
-        helix_angle, angle_multiplier, net = secondary_helix_force_kernel(
+        helix_angle, angle_multiplier, net = _secondary_helix_force_kernel(
             torque,
             spring_torque_breakdown.net,
             secondary_radius,
@@ -238,7 +285,7 @@ class PhysicalSecondaryPulley(SecondaryPulleyModel):
         """Calculate torsion spring torque (resists helix cam rotation)."""
         # Clamp shift distance to valid range
         # TODO: Remove
-        shift_distance = np.clip(shift_distance, 0, MAX_SHIFT)
+        shift_distance = np.clip(shift_distance, 0, self._max_supported_shift)
 
         # Calculate cam rotation from shift distance (approximation)
         # TODO: Improve relationship between shift and rotation
@@ -266,6 +313,7 @@ class PhysicalSecondaryPulley(SecondaryPulleyModel):
             Rotation angle [rad]
         """
         # Find x position that corresponds to this height
+        shift_distance = np.clip(shift_distance, 0, self._max_supported_shift)
         x_position = self.ramp.find_x_at_height(-shift_distance)
         # Get slope at that x position
         return x_position / HELIX_RADIUS
